feat(expensify): rename Update Report Status to Reimburse Report
The Integration Server report-status updater only supports marking an
APPROVED report as REIMBURSED (Approve/Reject return responseCode 410),
so the action now reflects that single purpose.

- Rename action expensify-update-report-status -> expensify-reimburse-report
- Remove the Status prop and hardcode REIMBURSED_STATUS in run()
- Update the action, Report ID, and summary descriptions accordingly
